restructer the app
While implementing react-google-maps/api I decided to break out that logic into RoutePlanner component. This is to insure a clean top level of state so in the future if I want to implement redux etc it will be easier. Also most pieces of state got renamed with more descriptive names. The form and DistanceService are working in this commit.